move emote serializer classes to emotesServer
